some more refactoring to get around the pesky cyclic imports.
