+++ DmpWorkflow/core/views.py
@@ -1,13 +1,15 @@
 import logging
 from copy import deepcopy
-from os.path import basename
+from datetime import datetime
 from json import loads, dumps
+from os.path import basename
+
 from flask import Blueprint, request, redirect, render_template, url_for
 from flask.ext.mongoengine.wtf import model_form
-from datetime import datetime
 from flask.views import MethodView
-from DmpWorkflow.core.DmpJob import DmpJob
+
 from DmpWorkflow.core.models import Job, JobInstance, HeartBeat, DataFile
+from DmpWorkflow.job.DmpJob import DmpJob
 
 jobs = Blueprint('jobs', __name__, template_folder='templates')
