+++ DmpWorkflow/scripts/dampe_cli_fetch_new_jobs.py
@@ -4,12 +4,15 @@ Created on Mar 15, 2016
 @author: zimmer
 """
 import logging
-from requests import get
-from sys import exit as sys_exit
 from argparse import ArgumentParser
-from DmpWorkflow.core.DmpJob import DmpJob
-from DmpWorkflow.config.defaults import DAMPE_WORKFLOW_URL, BATCH_DEFAULTS
 from importlib import import_module
+from sys import exit as sys_exit
+
+from requests import get
+
+from DmpWorkflow.config.defaults import DAMPE_WORKFLOW_URL, BATCH_DEFAULTS
+from DmpWorkflow.job.DmpJob import DmpJob
+
 HPC = import_module("DmpWorkflow.hpc.%s"%BATCH_DEFAULTS['system'])
 #from DmpWorkflow.scripts.dampe_cli_run_watchdog import __getRunningJobs
